Feature/CHO-102 - Frontend refactoring (#114)
* Refactoring of the frontend

* Refactoring of the frontend 2.0

* Add Dark Mode toggle and tooltip components

* Refactoring of the frontend 3.0

* Update contact page UI components and styles.- Refactor text area size and font size

* Update AboutPage and Team section UI styling

* fixed errors types

---------
