+++ messages/en.json
@@ -44,8 +44,8 @@
   "Change_Topic?": "Change Topic?",
   "Are_you_sure_you_want_to_change_the_topic?_This_will_lose_all_your_progress_in_the_current_session": "Are you sure you want to change the topic? This will lose all your progress in the current session.",
   "Dont_ask_me_again": "Don't ask me again",
-  "Log_in": "Log in",
-  "Log_out": "Log out",
+  "Log_in": "Log In",
+  "Log_out": "Log Out",
   "Enter_your_answer": "Enter your answer",
   "Congratulations": "Congratulations! 🎉",
   "You_have_completed_a_study_session": "You've completed a study session.",
@@ -74,6 +74,7 @@
   "Settings--": "Settings",
   "Above_is_the_prompt_we_use_to_generate_new_questions_for_you_It_is_not_editable": "Above is the prompt we use to generate new questions for you. It is not editable.",
   "Language": "Language",
+  "Dark_mode": "Dark Mode",
   "Feedback_Answer": "Feedback Answer",
   "Above_is_the_prompt_we_use_to_generate_hints_for_you_It_is_not_editable": "Above is the prompt we use to generate hints for you. It is not editable.",
   "Above_is_the_prompt_we_use_to_generate_answer_feedback_for_you_It_is_not_editable": "Above is the prompt we use to generate answer feedback for you. It is not editable",
