+++ messages/es.json
@@ -74,6 +74,7 @@
   "Settings--": "Configuraciones",
   "Above_is_the_prompt_we_use_to_generate_new_questions_for_you_It_is_not_editable": "Arriba está el prompt que usamos para generar nuevas preguntas para ti. No es editable.",
   "Language": "Idioma",
+  "Dark_mode": "Modo Oscuro",
   "Feedback_Answer": "Retroalimentación para la Respuesta",
   "Above_is_the_prompt_we_use_to_generate_hints_for_you_It_is_not_editable": "Arriba está el prompt que usamos para generar pistas para ti. No es editable.",
   "Above_is_the_prompt_we_use_to_generate_answer_feedback_for_you_It_is_not_editable": "Arriba está el prompt que usamos para generar retroalimentación de respuestas para ti. No es editable.",
