--- src/app/[locale]/(dashboard)/(routes)/notifications/page.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function NotificationsPage() {
-    return (
-        <div>NotificationsPage</div>
-    )
-}
