+++ src/app/[locale]/(dashboard)/(routes)/search/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function SearchPage() {
+    return (
+        <div>SearchPage</div>
+    )
+}
