+++ src/app/[locale]/posthog-page-view.tsx
@@ -1,4 +1,3 @@
-// app/PostHogPageView.tsx
 "use client";
 
 import { usePathname, useSearchParams } from "next/navigation";
@@ -6,21 +5,23 @@ import { useEffect } from "react";
 import { usePostHog } from "posthog-js/react";
 
 export default function PostHogPageView(): null {
-  const pathname = usePathname();
-  const searchParams = useSearchParams();
-  const posthog = usePostHog();
+  const pathname = usePathname(); // Get the current path from the URL
+  const searchParams = useSearchParams(); // Get the search parameters from the URL
+  const posthog = usePostHog(); // Get the PostHog instance from the context
+
   useEffect(() => {
-    // Track pageviews
+    // Capture a pageview event whenever the pathname or search parameters change
     if (pathname && posthog) {
-      let url = window.origin + pathname;
+      let url = window.origin + pathname; // Construct the current URL
       if (searchParams.toString()) {
-        url = url + `?${searchParams.toString()}`;
+        url = url + `?${searchParams.toString()}`; // Append search parameters if they exist
       }
+      // Capture the pageview event in PostHog
       posthog.capture("$pageview", {
         $current_url: url,
       });
     }
-  }, [pathname, searchParams, posthog]);
+  }, [pathname, searchParams, posthog]); // Dependencies for useEffect: run whenever these change
 
-  return null;
+  return null; // This component doesn't render anything
 }
