+++ src/components/footer.tsx
@@ -9,9 +9,7 @@ export default function Footer() {
   return (
     <footer className="flex flex-row justify-between items-center">
       <Badge>Beta</Badge>
-      <p className="text-xs text-gray-500">
-        {t("Chop_can_make_mistakes_Check_important_info")}
-      </p>
+
       {/* <Button
         onClick={() =>
           window.open("https://github.com/alvropena/chop-nextjs.git", "_blank")
