+++ src/components/header.tsx
@@ -1,40 +1,54 @@
-import React from "react";
+import Link from "next/link";
 import { Button } from "@/components/ui/button";
-import Logo from "@/components/logo";
-import { ModeToggle } from "@/components/mode-toggle";
-import { LogIn } from "lucide-react";
-import { useRouter } from "next/navigation";
-import { useUser } from "@auth0/nextjs-auth0/client";
-import { useTranslations } from "next-intl";
-import LanguageCombobox from "./language-combobox";
-import { AvatarDropdownMenu } from "./avatar-dropdown-menu";
-import SearchBarUser from "./search-bar-user";
+import { Sheet, SheetTrigger, SheetContent } from "@/components/ui/sheet";
+import { MenuIcon, MountainIcon, HomeIcon, UserIcon, SettingsIcon } from "lucide-react";
+import { UserAvatar } from "@/components/user-avatar";
 
-export default function Header() {
-  const { user } = useUser();
-  const router = useRouter();
-  const t = useTranslations("");
-
+export default function DashboardHeader() {
   return (
-    <header className="flex flex-row items-center justify-between">
-      <Logo />
-      <div className="flex flex-row items-center gap-4">
-        <SearchBarUser />
-        <LanguageCombobox />
-        <ModeToggle />
-        {!user ? (
-          <Button
-            className="gap-2"
-            onClick={() => {
-              router.push("/api/auth/login");
-            }}
-          >
-            <LogIn className="h-4 w-4" />
-            {t("Log_in")}
+    <header className="sticky top-0 z-30 flex h-14 items-center gap-4 border-b bg-background px-4 md:static md:h-auto md:border-0 md:bg-transparent md:px-6">
+      <Sheet>
+        <SheetTrigger asChild>
+          <Button size="icon" variant="outline" className="md:hidden">
+            <MenuIcon className="h-5 w-5" />
+            <span className="sr-only">Toggle Menu</span>
           </Button>
-        ) : (
-          <AvatarDropdownMenu />
-        )}
+        </SheetTrigger>
+        <SheetContent side="left" className="md:max-w-xs">
+          <nav className="grid gap-6 text-lg font-medium">
+            <Link
+              href="/"
+              className="group flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-primary text-lg font-semibold text-primary-foreground"
+              prefetch={false}
+            >
+              <MountainIcon className="h-5 w-5 transition-all group-hover:scale-110" />
+              <span>Chop Inc</span>
+            </Link>
+            <Link
+              href="/home"
+              className="flex items-center gap-4 px-2.5 text-muted-foreground hover:text-foreground"
+              prefetch={false}
+            >
+              <HomeIcon className="h-5 w-5" />
+              Home
+            </Link>
+            <Link href="/profile" className="flex items-center gap-4 px-2.5 text-foreground" prefetch={false}>
+              <UserIcon className="h-5 w-5" />
+              Profile
+            </Link>
+            <Link
+              href="/settings"
+              className="flex items-center gap-4 px-2.5 text-muted-foreground hover:text-foreground"
+              prefetch={false}
+            >
+              <SettingsIcon className="h-5 w-5" />
+              Settings
+            </Link>
+          </nav>
+        </SheetContent>
+      </Sheet>
+      <div className="ml-auto flex items-center gap-2">
+        <UserAvatar />
       </div>
     </header>
   );
