+++ src/components/landing-footer.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function LandingFooter() {
+    return (
+        <footer className="flex flex-row items-center justify-center w-full p-6 border-t-2">
+            <p className="text-sm">&copy; 2024 Chop Inc. All rights reserved.</p>
+        </footer>
+    )
+}
