+++ src/components/logo.tsx
@@ -1,20 +1,15 @@
-import { useTheme } from "next-themes";
+import React from 'react'
 import Image from "next/image";
 import Link from "next/link";
 
-import React from 'react'
-
 export default function Logo() {
-    const { theme } = useTheme()
-    const logoSrc = theme === "dark" ? "/logo.svg" : "/logo_d.svg";
-
     return (
-        <Link href={"/"}>
+        <Link href={"/home"}>
             <Image
                 height={80}
                 width={80}
                 alt="logo"
-                src={logoSrc}
+                src={"/logo_d.svg"}
             />
         </Link>
     )
