--- src/components/sidebar/sidebar-item.tsx
@@ -1,30 +1,0 @@
-"use client";
-
-import { Button } from "@/components/ui/button";
-import { usePathname, useRouter } from "next/navigation";
-import { cn } from "@/lib/utils";
-
-interface SidebarItemProps {
-    icon: any;
-    label: string;
-    href: string;
-}
-
-export const SidebarItem = ({
-    label, href, icon
-}: SidebarItemProps) => {
-    const pathname = usePathname();
-    const router = useRouter();
-
-    const isActive = (pathname === "/" && href === "/") || pathname === href || pathname?.startsWith(`${href}/`);
-
-    const onClick = () => {
-        router.push(href);
-    }
-
-    return (
-        <Button onClick={onClick}>
-            {label}
-        </Button>
-    )
-}
