+++ src/lib/typing-effect.tsx
@@ -1,4 +1,3 @@
-// typing-effect.tsx
 "use client";
 
 import React, { useState, useEffect } from 'react';
